% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage of missing mutations to a data set}
\usage{
add_coverage(coco, coverage, mutation_list)
}
\arguments{
\item{coco}{The data set.}

\item{coverage}{The coverage filen with position as the first column and coverage as the second.}

\item{mutation_list}{The mutations to be added (usually \code{colnames(lineage_defs)}).}
}
\value{
coco, but with added rows.
}
\description{
Add coverage of missing mutations to a data set
}
