% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcov.R
\name{alcov}
\alias{alcov}
\title{Simplified Alcov Model Using Non-Negative Linear Regression}
\usage{
alcov(Y, lmps, muts)
}
\arguments{
\item{Y}{Vector of frequencies for each sample.}

\item{lmps}{Matrix or data frame of lineage definitions, similar to varmat.}

\item{muts}{Char vector of mutation names, used to select and order columns in lmps if it is a data frame.}
}
\value{
An "Alcov" object with coefficients from the linear regression model, representing the estimated proportions of variants. Object can be plotted.
}
\description{
A simplified Alcov model for linear regression without intercepts,
ensuring positivity of the coefficients. This function scales inputs and
applies a non-negative linear regression to estimate variant proportions.
}
\examples{
Y <- c(0.1, 0.2, 0.3, 0.4)
lmps <- matrix(c(0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1), nrow = 4, byrow = TRUE)
muts <- c("mut1", "mut2", "mut3")

coef <- alcov(Y, lmps, muts)
print(coef)

If you want to plot the coefficients:
plot(coef, muts)

}
