% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R
\name{astronomize}
\alias{astronomize}
\title{Extract mutation list from a directory of constellation files.}
\usage{
astronomize(path = NULL)
}
\arguments{
\item{path}{Path to the constellations folder in the cov-lineages/constellations repository. The default assumes that the current project and the constellations repo are in the same directory.}
}
\value{
A variant matrix for use with provoc.
}
\description{
"Constellations" are files produced from \url{https://github.com/cov-lineages/constellations}, which represent the cov-lineage team's best knowledge about which mutations define a lineage. These are updated frequently, please \code{git pull} in your cloned copy accordingly. See details.
}
\details{
Code is adapted from \code{scripts/estimate-freqs.R} in \url{https://github.com/PoonLab/gromstole}.

From the repo, a constellation "a collection of mutations which are functionally meaningful, but which may arise independently a number of times".

This function requires a local clone of the constellations repository. By default, I assume that your current project is in the same parent directory as the constellation file (e.g. your working directory is \code{.../parent/current_project} and the constellations repo is cloned as \code{.../parent/constellations}).

If this is not the case, a path to the root folder of the constellations repo is required. The path is to the root folder, not the folder with the constellation files.

There are no options to specify which lineages to include. I am working on a \code{annihilate(coco, varmat)} function to remove mutations that aren't shared between coco and varmat and squash lineages that have too few observed mutations or are too similar to other lineages.
}
\examples{
if(dir.exists("../constellations")) {
    varmat <- astronomize()
}

}
