% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{autoplot.provoc}
\alias{autoplot.provoc}
\title{Plot the results using ggplot2}
\usage{
\method{autoplot}{provoc}(provoc_obj, date_col = NULL)
}
\description{
Only plots the results, does not plot residuals
}
