% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_binom.R
\name{coda_binom}
\alias{coda_binom}
\title{Estimate a Bayesian model for rho.}
\usage{
coda_binom(
  coco,
  varmat,
  prm = list(adapt = 500, burnin = 1000, sample = 1000, thin = 4, quiet = 0)
)
}
\arguments{
\item{coco}{A data frame containing counts, coverage, and mutation names}

\item{varmat}{The variant matrix to be used in the study. The rownames must be the VoCs and the colnames must be the mutation names (in the same format as the mutation names in \code{coco})}

\item{prm}{The parameters for the mcmc fit, generally from a prm.json file.}
}
\value{
an mcmc.list object with each column representing the proportion of the variant of concern, in the order of the rownames of variantmat. It is suggested to use melt_mcmc to get output that plays better with ggplot2 and dplyr.
}
\description{
Estimate a Bayesian model for rho.
}
