% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations_by_lineage.R
\docType{data}
\name{constellation_lists}
\alias{constellation_lists}
\title{Lists of mutations per lineage, as per Constellations}
\format{
A data frame
}
\source{
\url{https://github.com/cov-lineages/constellations}
}
\usage{
constellation_lists
}
\description{
The names of mutations according to Pangolin's "Constellations" files. These are current as of October 2024. The constellation files for pre-Omicron variants are not expected to change.
}
\keyword{datasets}
