% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{convergence}
\alias{convergence}
\title{Check if provoc converged}
\usage{
convergence(res, verbose = TRUE)
}
\arguments{
\item{res}{The result of \code{provoc()}}

\item{verbose}{Print a message to the screen?}
}
\value{
Invisbly returns TRUE if all samples converged, false otherwise.
}
\description{
Check if provoc converged
}
