% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copt_binom.R
\name{copt_binom}
\alias{copt_binom}
\title{Estimate the proportions of VOCs using Constrained OPTimization}
\usage{
copt_binom(coco, varmat)
}
\arguments{
\item{coco}{A data frame containing counts, coverage, and mutation names}

\item{varmat}{The variant matrix to be used in the study. The rownames must be the VoCs and the colnames must be the mutation names (in the same format as the mutation names in \code{coco})}
}
\value{
A list containing the vector of proportions for the variants of concern, in the same order as the rows of varmat. This list also contains valuable convergence information.
}
\description{
Estimate the proportions of VOCs using Constrained OPTimization
}
\details{
The estimates are found by minimizing the squared difference between the frequency of each mutation and the prediction of a binomial model where the proportion is equal to the sum of rho times the relevant column of varmat and the size parameter is equal to the coverage.

The algorithm will first try a prior guess based on the current (March 2022) most common VOCs, then will try a uniform proportion, then (the nuclear option) will try 100 random perturbations until it works. Fails gracefully, with list elements indicating the convergence status and the initialization of rho.

This function currently does not return any estimate of variance for the proportions and should not be trusted beyond a quick check.
}
