% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R
\name{covariants}
\alias{covariants}
\title{Gather CoVariants defintions}
\usage{
covariants(
  filename = "covariants.csv",
  path = c("data/clean/", "data/", "./")[1],
  update = FALSE,
  url =
    "https://raw.githubusercontent.com/hodcroftlab/covariants/master/scripts/clusters.py"
)
}
\arguments{
\item{filename}{Save to disk so that the python needn't be run every time. If file exists, it is loaded rather than re-downloaded.}

\item{path}{Path to which to read or write file.}

\item{update}{If TRUE, downloads the data no matter what.}

\item{url}{Useful if the url ever changes.}
}
\description{
CoVariants are stored as a python script, so this code downloads the file, runs it via \code{reticulate}, then parses the output.
}
