% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage_at_aa}
\alias{coverage_at_aa}
\title{Get coverage at mutation positions.}
\usage{
coverage_at_aa(coverage, aa)
}
\arguments{
\item{coverage}{A data frame with columns labelled position (1-indexed) and coverage.}

\item{aa}{A vector of mutations. aa:orf1a:I300V, ins:28215:3, del:27378:25, or C703T}
}
\description{
Currently requires "aa" notation.
}
