% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_diagnostics.R
\name{coverage_by_lineage_defs}
\alias{coverage_by_lineage_defs}
\title{Summarise coverage by lineage/mutation combo}
\usage{
coverage_by_lineage_defs(provoc_obj, fun = mean, ...)
}
\arguments{
\item{provoc_obj}{Result of \code{provoc()}.}

\item{fun}{The function used to summarise the coverage information.}

\item{...}{Further arguments passed onto \code{fun()}, such as \code{prob = 0.5} to find the median when fun = \code{quantile}.}
}
\value{
A matrix with the total column and row names of all lineage defs, and entries indicating the number of analyses which used that mutation in that lineage.
}
\description{
Useful for finding whether some calls were based on less information
}
