% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{delete.spectra.h5}
\alias{delete.spectra.h5}
\title{delete a couple of spectra in a h5 file}
\usage{
delete.spectra.h5(ID_h5 = 1, num = 1, w_d = getwd())
}
\arguments{
\item{ID_h5}{the ID give by info.h5()}

\item{num}{the number correspoding at the spectra corrupted}

\item{w_d}{working directory}
}
\value{
a modifiy h5 file.
}
\description{
sometimes, an experience is stopped brutaly. The last spectra is corrupted.
The file can be reading by the import.h5() function. info.h5 and delete.spectra.h5()
can be used for resolving the problem.
}
\examples{
# Use the info.h5() function before use that. And read the help for understand why
# we deleted the spectra number 176 of ID6.

# delete.spectra.h5(6,176)

# After that, you have a new h5 file (mod_A3_40.h5). You can exhile the original file.
}
