% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{differ_by_one_or_less}
\alias{differ_by_one_or_less}
\title{Finds if two vectors only differ between one mutation}
\usage{
differ_by_one_or_less(v1, v2)
}
\arguments{
\item{v1}{vector for comparison}

\item{v2}{vector for comparison}
}
\value{
TRUE, if they only differ by one mutation
}
\description{
Finds if two vectors only differ between one mutation
}
