% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{dy.spectra}
\alias{dy.spectra}
\title{Plot spectra dynamic.}
\usage{
dy.spectra(
  sel_sp = sp$mt$meta[sp$acq, "end"],
  L = sp,
  new_color = FALSE,
  name = FALSE
)
}
\arguments{
\item{sel_sp}{a numeric vector with the selection of spectra.}

\item{L}{sp}

\item{new_color}{Logical TRUE/FALSE}

\item{name}{the title of plot and the name of the file.}
}
\value{
a html plot
}
\description{
Export a html plot for an intuitive exploration of data. The number of spectra is limited at 60
for increase the comfort of viewing.
}
\examples{
# dy.spectra(sp$mt$meta[sp$acq,"end"], new_color = FALSE)
# dy.spectra(1:50)
# dy.spectra(c(1, 4, 8, 22, 30), new_color = TRUE)
}
