% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{empty.meta}
\alias{empty.meta}
\title{Export a meta file
Execute this function for create a initial file "meta_empty.csv".}
\usage{
empty.meta(L = sp)
}
\arguments{
\item{L}{sp}
}
\value{
sp
}
\description{
Export a meta file
Execute this function for create a initial file "meta_empty.csv".
}
\examples{
# empty.meta()
}
