% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{export.info}
\alias{export.info}
\title{extrat info of h5 file}
\usage{
export.info(num = 1, w_dir = w_d)
}
\arguments{
\item{num}{a numeric vector}

\item{w_dir}{working directory}
}
\value{
a list with informations
}
\description{
extrat info of h5 file
}
