% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{extant_lineages}
\alias{extant_lineages}
\title{Filter lineages active on a given date.}
\usage{
extant_lineages(
  lineage_names,
  start_date,
  check_after = TRUE,
  check_canada = FALSE
)
}
\arguments{
\item{lineage_names}{A character vector of lineage names. Must match the names in \code{mutations_by_lineage}.}

\item{start_date}{Earliest date in the study. Must be in ISO-8601 format (as should all dates with no exceptions).}

\item{check_after}{Check if there's an observed sequence after the start of your study? Set FALSE if the start date is recent.}

\item{check_canada}{Checks if the lineage was ever observed in Canada. Default FALSE.}
}
\value{
A character vector.
}
\description{
Using NGSB data (see \code{?mutations_by_lineage}), checks that the earliest sequence in a lineage was observed by \code{start_date}. Optionally checks if the latest sequence in the lineage was observed after start_date. Optionally checks if the lineage was ever observed in Canada.
}
\details{
The code also ignores any + symbols and anything after them, so lineages such as B.1.617.2+K417N (Delta+) will be treated as B.1.617.2 (Delta).
}
