% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R
\name{filter_lineages}
\alias{filter_lineages}
\title{Filter linmat for specific lineages, keeping mutations that are present in at least one lineage}
\usage{
filter_lineages(
  linmat = NULL,
  lineages = c("B.1.526", "B.1.1.7", "B.1.351", "B.1.617.2", "B.1.427", "B.1.429", "P.1"),
  return_df = FALSE,
  path = NULL,
  shared_order = TRUE
)
}
\arguments{
\item{linmat}{The result of \code{astronomize()}. If NULL, tries to run \code{astronoimize}.}

\item{lineages}{Vector of lineage names (must be in \code{rownmaes(linmat)}). Defaults to lineages circulating in 2021-2022.}

\item{return_df}{Should the function return a data frame? Note that returned df is transposed compared to linmat. Default FALSE.}

\item{path}{Passed on to \code{astronomize} if \code{linmat} is NULL.}

\item{shared_order}{Put shared mutations first? Default TRUE.}
}
\value{
A lineage matrix with fewer rows and columns than \code{linmat}. If \code{return_df}, the columns represent lineage names and a \code{mutations} column is added.
}
\description{
Filter linmat for specific lineages, keeping mutations that are present in at least one lineage
}
\details{
After removing some lineage, the remaining mutations might not be present in any of the remaining lineage. This function will remove mutations that no longer belong to any lineage.

Note that return_df will
}
