% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R
\name{filter_variants}
\alias{filter_variants}
\title{Filter varmat for specific variants, keeping mutations that are present in at least one variant}
\usage{
filter_variants(
  variants = c("B.1.526", "B.1.1.7", "B.1.351", "B.1.617.2", "B.1.427", "B.1.429", "P.1"),
  varmat = NULL,
  return_df = FALSE,
  path = NULL,
  shared_order = TRUE
)
}
\arguments{
\item{variants}{Vector of variant names (must be in \code{rownmaes(varmat)}). Defaults to variants circulating in 2021-2022.}

\item{varmat}{The result of \code{astronomize()}. If NULL, tries to run \code{astronoimize}.}

\item{return_df}{Should the function return a data frame? Note that returned df is transposed compared to varmat. Default FALSE.}

\item{path}{Passed on to \code{astronomize} if \code{varmat} is NULL.}

\item{shared_order}{Put shared mutations first? Default TRUE.}
}
\value{
A variant matrix with fewer rows and columns than \code{varmat}. If \code{return_df}, the columns represent variant names and a \code{mutations} column is added.
}
\description{
Filter varmat for specific variants, keeping mutations that are present in at least one variant
}
\details{
After removing some variants, the remaining mutations might not be present in any of the remaining variants. This function will remove mutations that no longer belong to any variants.

Note that return_df will
}
