% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{find_mutation_column}
\alias{find_mutation_column}
\title{Finds the right mutation column in lineage_defs.}
\usage{
find_mutation_column(data, lineage_defs)
}
\arguments{
\item{data}{Data frame containing count, coverage, and lineage columns}

\item{lineage_defs}{Optional mutation definitions}
}
\value{
A list of length 2 containing the possibly transposed lineage_defs and
the right mutation column.
}
\description{
Checks the columns of lineage_defs and if not found in columns will check the rows
and transpose lineage_defs.
}
