% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freyja.R
\name{freyja}
\alias{freyja}
\title{Simplified Freyja Model Using Lasso Regression}
\usage{
freyja(mix, depths, df_barcodes, muts, eps = 1e-04)
}
\arguments{
\item{mix}{Vector of frequencies (count divided by coverage) for each sample.}

\item{depths}{Vector of coverage for each sample.}

\item{df_barcodes}{Numeric matrix varmat, with rows as samples and columns as mutations.}

\item{muts}{Char vector of mutation names.}

\item{eps}{Very small number representing the strength for Lasso regression. Default is 1e-4.}
}
\value{
A "Freyja" object with coefficients from the Lasso regression, representing the estimated proportions of variants. Object can be plotted.
}
\description{
This function applies a Lasso regression to estimate proportions of variants
based on frequency and adjusting for sequence depth.
}
\examples{
mix <- c(0.1, 0.2, 0.3, 0.4)
depths <- c(10, 20, 30, 40)
df_barcodes <- matrix(c(0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1), nrow = 4, byrow = TRUE)
muts <- c("mut1", "mut2", "mut3")

coef <- freyja(mix, depths, df_barcodes, muts)
print(coef)

If you want to plot the coefficients:
plot(coef, muts)

}
