% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{fx.spectra}
\alias{fx.spectra}
\title{Title}
\usage{
fx.spectra(
  sel_sp = sp$mt$meta[sp$acq, "end"],
  pkm = 59,
  pkM = 205,
  L = sp,
  new_title = "fx_spectra",
  new_color = FALSE,
  leg = "r"
)
}
\arguments{
\item{leg}{}
}
\value{

}
\description{
Title
}
