% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mutations.R
\docType{data}
\name{gcode}
\alias{gcode}
\title{Amino acid one-letter codes}
\format{
An object of class \code{list} of length 66.
}
\usage{
gcode
}
\description{
As with many things in this repo, lovingly stolen from \url{https://github.com/PoonLab/gromstole}.
}
\keyword{datasets}
