% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{get_actual_defs}
\alias{get_actual_defs}
\title{Extract the actual lineage defitions used when fitting the models.}
\usage{
get_actual_defs(provoc_obj)
}
\description{
Returns a list if \code{by_col} was specified.
}
