% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{get_lineage_defs}
\alias{get_lineage_defs}
\title{Extract the mutation definitions used to fit the model}
\usage{
get_lineage_defs(provoc_obj)
}
\arguments{
\item{provoc_obj}{Object of class provoc to be used to extract the mutation definitions.}
}
\description{
Extract the mutation definitions used to fit the model
}
