% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineage_info.R
\name{get_lineage_info}
\alias{get_lineage_info}
\title{Gather information about lineages}
\usage{
get_lineage_info(path = "data")
}
\arguments{
\item{path}{Path to save (or load) the lineage information. Default \code{data/}.}
}
\value{
A data frame, with the side effect of creating a file \code{path/lineage-list.csv} if it does not exist.
}
\description{
Downloads and parses the table at \url{https://cov-lineages.com/lineages.org} using the \code{rvest} package.
}
