% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{hprint}
\alias{hprint}
\title{print the text below by the hour}
\usage{
hprint(txt = "hello there")
}
\arguments{
\item{txt}{a character string}
}
\value{
a character string more the hour
}
\description{
print the text below by the hour
}
\examples{
# hprint(txt = "hello there")
}
