% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{ind.acq}
\alias{ind.acq}
\title{return index of spectra for each acquistion}
\usage{
ind.acq(n_acq, L)
}
\arguments{
\item{n_acq}{a number}

\item{L}{sp}
}
\value{
a numeric vector
}
\description{
return index of spectra for each acquistion
}
\examples{
# For just one acquisition :
# ind.acq(1,sp)
# Or for more :
# sapply(1:10,ind.acq,L=sp,simplify = TRUE)
}
