% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{ind.pk}
\alias{ind.pk}
\title{return index of peaks}
\usage{
ind.pk(ms, mode = c("max", "exact", "all"), L = sp)
}
\arguments{
\item{ms}{the mass of the peak}

\item{mode}{a character. Specifies whether the returned index is the exact peak (exact), the maximum peak (max) or all peaks (all) related to a unit mass.}

\item{L}{sp}
}
\value{
a numeric vector
}
\description{
return index of peaks
}
\examples{
# For get the index for one peak :
# ind.pk(137)
# [1] 318
# Or for more :
# sapply(c(81,137,205),ind.pk)
}
