% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{info.h5}
\alias{info.h5}
\title{summarise the h5 files}
\usage{
info.h5(w_d = getwd())
}
\arguments{
\item{w_d}{working directory}
}
\value{
a list with informations
}
\description{
the function return a list of two objects, one matrix and one list. The summary show all acquisitions in column
and others informations in row. Theses informations are the ID, the acquisition time (in second),
the length of x mass, the number of spectra, index of start and end, and
the smallest value among the maximum intensities of each spectrum of the acquisition.
}
\details{
the other list is order by ID. Each table includes the average, median and
maximum intensities of each spectrum of the acquisition.
}
\examples{
# l_info <- info.h5()
# l_info$summary

# A2_20  A2_30  A2_40  A3_20  A3_30  A3_40
# ID                1      2      3      4      5      6
# acq time (s)     10     10     10     10     10      0
# length abs   158768 158768 158768 158768 158768 158768
# nbr spectra     180    180    180    180    180    176
# start             1    181    361    541    721    901
# end             180    360    540    720    900   1076
# min Imax     184083 112189 251471  47338 137394      0

# We can view than the A3_40, ID 6, have an intensity at zero. And more, all
# acquisitions are 180 spectra and acquisition during 10 seconds except the ID 6.

# l_info$$byID[[6]]
#         A3_40_174    A3_40_175     A3_40_176
# mean    407.22151    404.94106             0
# median  13.36089     13.32512              0
# max     792628.75    789557.87             0

# The last spectra is corrompted. We should deleted the number 176.

}
