% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{is_subset}
\alias{is_subset}
\title{Finds if one lineage is a subset of another}
\usage{
is_subset(v1, v2)
}
\arguments{
\item{v1}{vector for comparison}

\item{v2}{vector for comparison}
}
\value{
Result, TRUE if v2 is a subset of v1
}
\description{
Finds if one lineage is a subset of another
}
