% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{jaccard_simularity}
\alias{jaccard_simularity}
\title{Finds the Jaccard similarity between two vectors}
\usage{
jaccard_simularity(v1, v2)
}
\arguments{
\item{v1}{vector for comparison}

\item{v2}{vector for comparison}
}
\value{
The Jaccard simularity
}
\description{
Finds the Jaccard similarity between two vectors
}
