% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_diagnostics.R
\name{left_both_right}
\alias{left_both_right}
\title{Calculate the set differences for two character vectors}
\usage{
left_both_right(left, right)
}
\arguments{
\item{left, right}{Two character vectors}
}
\description{
Calculate the set differences for two character vectors
}
\keyword{internal}
