% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{lineage_de_from_data}
\alias{lineage_de_from_data}
\title{Use mutations found in the data to determine the lineage matrix}
\usage{
lineage_de_from_data(
  type = NULL,
  pos = NULL,
  alt = NULL,
  aa = NULL,
  data = NULL,
  mutation_format = c("tpa", "aa"),
  max_n = 80,
  top_quantile = 0.05,
  matches = 3,
  start_date = NULL,
  check_after = TRUE,
  check_canada = FALSE,
  mutations = provoc::mutations_by_lineage
)
}
\arguments{
\item{type}{A vector containing either +, -, or ~, representing insertions, deletions, and mutations, respectively. Can also be i, d, or m.}

\item{pos}{A vector containing the 1-indexed positions.}

\item{alt}{A vector containing the alternative (nucleotide(s) for mutations and insertions or the number of deletions).}

\item{aa}{The vector of mutations in amino acid format (like those create by \code{parse_mutation}). TODO: Not yet implemented.}

\item{data}{A data frame containing columns labelled "type", "pos", and "alt", each with entries in the expected format. TODO: Not yet implemented.}

\item{mutation_format}{"tpa" (default) for \code{type|pos|alt} (e.g. \code{~|2832|G}; note that it includes the pipes "|") or "aa" for amino acid (e.g. \code{aa:orf1a:K856R}).}

\item{max_n}{The maximum number of mutations from a given lineage to retain.}

\item{top_quantile}{Only take mutations that are in the top \code{top_quantile} quantile. E.g. 0.95 gives the mutations that are observed in more than 95\% of the sequences of that lineage.}

\item{matches}{The minimum number of matching mutations in the NextStrain GenBank data.}

\item{start_date}{Earliest date in the study. Must be in ISO-8601 format (as should all dates with no exceptions).}

\item{check_after}{Check if there's an observed sequence after the start of your study? Set FALSE if the start date is recent.}

\item{check_canada}{Checks if the lineage was ever observed in Canada. Default FALSE.}

\item{lineage_names}{A character vector of lineage names. Must match the names in \code{mutations_by_lineage}.}
}
\value{
A lineage matrix (rownames are lineage, colnames are mutations, entry i,j is 1 if lineage i contains mutation j, 0 otherwise).
}
\description{
Uses the included GenBank mutation list (\code{?mutations_by_lineage}) to determine a lineage matrix that is guaranteed to have the mutations from your data (assuming you provide the correct mutation name format).
}
