% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{lineage_defs_from_lineage}
\alias{lineage_defs_from_lineage}
\title{Given a vector of lineage names, uses GenBank data to determine the lineage matrix.}
\usage{
lineage_defs_from_lineage(
  lineages,
  max_n = NULL,
  top_quantile = NULL,
  mutation_format = c("tpa", "aa"),
  mutations = provoc::mutations_by_lineage
)
}
\arguments{
\item{lineages}{A character vector of lineage names in the same format as the}

\item{max_n}{The maximum number of mutations from a given lineage to retain.}

\item{top_quantile}{Only take mutations that are in the top \code{top_quantile} quantile. E.g. 0.05 gives the mutations in more than 95\% of the sequences of that lineage.}

\item{mutation_format}{"tpa" (default) for \code{type|pos|alt} (e.g. \code{~|2832|G}; note that it includes the pipes "|") or "aa" for amino acid (e.g. \code{aa:orf1a:K856R}).}
}
\value{
A lineage matrix (rownames are lineage, colnames are mutations, entry i,j is 1 if lineage i contains mutation j, 0 otherwise).
}
\description{
Requires one of a few potentially arbitrary choices: Use the \code{max_n} most common mutations for a given lineage or use the mutations in the top \code{top_quantile} quantile.
}
\details{
See \code{?mutations_by_lineage} for more information about the GenBank data used and see \code{unique(mutations_by_lineage$lineage)} to see available lineage names.
}
