% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{lineage_defs_from_list}
\alias{lineage_defs_from_list}
\title{Create a lineage definition matrix from a list of lineages with their mutations}
\usage{
lineage_defs_from_list(lineage_list)
}
\arguments{
\item{lineage_list}{A named list of vectors of mutations. Mutation names should be in the same format as those in your data, otherwise post-processing will be required.}
}
\value{
A lineage matrix (rownames are lineages, colnames are mutations, entry i,j is 1 if lineage i contains mutation j, 0 otherwise).
}
\description{
The input should be a named list where the names represent lineages and the values are vectors of mutations. The output will be a valid lineage matrix.
}
