% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lineage_facts}
\alias{lineage_facts}
\title{Facts about the lineages}
\format{
A data frame
}
\source{
\url{https://nextstrain.org}
}
\usage{
lineage_facts
}
\description{
Especially useful for checking the earliest/latest sequence date
}
\keyword{datasets}
