% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{M.Z.max}
\alias{M.Z.max}
\title{search the highest peak in accord to a mass number}
\usage{
M.Z.max(ma, L = sp)
}
\arguments{
\item{ma}{a number of mass}

\item{L}{sp}
}
\value{
a numeric vector with all the exact mass closed to the mass number
}
\description{
search the highest peak in accord to a mass number
}
\examples{
# M.Z.max(c(59, 137))
# 59.233 137.121
}
