% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{M.Z}
\alias{M.Z}
\title{search all peak in accord to a mass number}
\usage{
M.Z(ma, L = sp)
}
\arguments{
\item{ma}{a number of mass}

\item{L}{sp}
}
\value{
a numeric vector with all the exact mass closed to the mass number

a numeric vector with all the exact mass closed to the mass number
}
\description{
search all peak in accord to a mass number
}
\examples{
# M.Z(c(59, 137))
# 58.873  59.045  59.233  59.267  59.320  59.405 137.037 137.121
}
