% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{mcr.voc}
\alias{mcr.voc}
\title{mcr.voc is a large function to easily do an MCR analysis for the dataset.
As output, you can retrieve figures and MCR results.}
\usage{
mcr.voc(
  ncMCR = 3,
  grp = NULL,
  pk_sel = "all",
  time_format = c("date", "time"),
  Li = sp
)
}
\arguments{
\item{ncMCR}{integer; number of componant of MCR}

\item{grp}{a character string of the group 's column name}

\item{pk_sel}{a vector of selected peaks, or "all"}

\item{time_format}{a charater string "date" or "time"}

\item{Li}{the list with spectra (sp)}
}
\value{
return figures and results
}
\description{
mcr.voc is a large function to easily do an MCR analysis for the dataset.
As output, you can retrieve figures and MCR results.
}
\examples{
# mcr.result <- mcr.voc(ncMCR = 3, grp = "modality")
}
