% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{meta.ctrl}
\alias{meta.ctrl}
\title{print a lot of graph for a rapid overview of metadata kinetic}
\usage{
meta.ctrl(L = sp, short = TRUE)
}
\arguments{
\item{L}{sp}

\item{short}{a short selection of meta. FALSE for all graph}
}
\value{
a plot
}
\description{
print a lot of graph for a rapid overview of metadata kinetic
}
\examples{
# meta.ctrl()
}
