% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{monitor_plot_AUC}
\alias{monitor_plot_AUC}
\title{Title}
\usage{
monitor_plot_AUC(
  M_num = M.Z(c(59, 137)),
  each_mass = TRUE,
  group = FALSE,
  graph_type = "dy",
  L = sp,
  Y_exp = FALSE,
  time_format = "date"
)
}
\arguments{
\item{time_format}{}
}
\value{

}
\description{
Title
}
