% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R, R/parse_mutations.R
\docType{data}
\name{orfs}
\alias{orfs}
\title{List of open reading frames}
\format{
An object of class \code{list} of length 12.

An object of class \code{list} of length 12.
}
\usage{
orfs

orfs
}
\description{
As with many things in this repo, lovingly stolen from \url{https://github.com/PoonLab/gromstole}.
}
\keyword{datasets}
