% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mutations.R
\name{parse_ivar}
\alias{parse_ivar}
\title{Parse Mutations from the ivar pipeline}
\usage{
parse_ivar(ivar_output)
}
\arguments{
\item{ivar_output}{A data frame imported from the ivar output. Must have columns POS and ALT}
}
\description{
Parse Mutations from the ivar pipeline
}
