% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mutations.R
\name{parse_mutation}
\alias{parse_mutation}
\title{Parse mutations from (type, pos, alt) to amino acid}
\usage{
parse_mutation(
  type,
  pos,
  alt,
  reffile = system.file("extdata/NC_045512.fa", package = "provoc")
)
}
\arguments{
\item{type}{Either "~", "+", or "-" for mutation, insertion, or deletion, respectively.}

\item{pos}{The one-indexed position relative to the reference.}

\item{alt}{The alternate nucleotide, nucleotides, or number of deletions, respectively.}

\item{reffile}{The path to the reference file that was used to define mutations. Usually NC_045512.fa except in very particular circumstances.}
}
\value{
e.g. aa:orf1a:K856R
}
\description{
As with many things in this repo, lovingly stolen from the \code{seq_utils.py} script in \url{https://github.com/PoonLab/gromstole} (as well as PoonLab/covizu).
}
