% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mutations.R
\name{parse_mutations}
\alias{parse_mutations}
\title{Parse output of the Gromstole pipeline, from SNVs to AAs}
\usage{
parse_mutations(labels)
}
\arguments{
\item{labels}{The \code{labels} column from GromStole output.}
}
\value{
A vector of the same length of \code{labels}.
}
\description{
This function can be slow.
}
