% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mutations.R
\name{parse_one_ivar}
\alias{parse_one_ivar}
\title{Parse mutations from ivar pipeline}
\usage{
parse_one_ivar(pos, alt)
}
\arguments{
\item{pos}{The position on the genome}

\item{alt}{The alternate genotype}
}
\description{
Parse mutations from ivar pipeline
}
