% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mutations.R
\name{parse_unique_mutations}
\alias{parse_unique_mutations}
\title{Parse all unique mutations in a vector}
\usage{
parse_unique_mutations(muts)
}
\arguments{
\item{muts}{A vector of mutations in the format "+50535C", "-43234.2", and "+342234.AC".}
}
\value{
A data frame with columns \code{label} and \code{mutation}.
}
\description{
Parse all unique mutations in a vector
}
