% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcov.R
\name{plot.alcov}
\alias{plot.alcov}
\title{Plot Method for Alcov Coefficients}
\usage{
\method{plot}{alcov}(coef, muts)
}
\arguments{
\item{coef}{Vector of coefficients returned by the alcov function.}

\item{muts}{Vector of mutation names, which must match the length of \code{coef}.}
}
\description{
Plot Method for Alcov Coefficients
}
