% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freyja.R
\name{plot.freyja}
\alias{plot.freyja}
\title{Plot Method for Freyja Coefficients}
\usage{
\method{plot}{freyja}(coef, muts)
}
\arguments{
\item{coef}{Vector of coefficients returned by the freyja function.}

\item{muts}{Vector of mutation names, which must match the length of \code{coef}.}
}
\description{
Plot Method for Freyja Coefficients
}
