% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{plot.provoc}
\alias{plot.provoc}
\title{Plot the results of model fitting}
\usage{
\method{plot}{provoc}(provoc_obj, plot_type = c("barplot"))
}
\arguments{
\item{provoc_obj}{Resulting object of class provoc to be used in plotting.}

\item{plot_type}{Currently only "barplot" is implemented. Residual plots and other diagnostics are works in progress.}
}
\description{
Prints the most useful diagnostics to the screen, invisibly returning them as a list.
}
