% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_diagnostics.R
\name{plot_actual_defs}
\alias{plot_actual_defs}
\title{Visualize how lineage definitions were used in analysis}
\usage{
plot_actual_defs(
  provoc_obj,
  type = "coverage",
  fun = mean,
  ...,
  main = NULL,
  col = hcl.colors(n = 21, palette = "Dark Mint", rev = TRUE)
)
}
\arguments{
\item{provoc_obj}{The result of calling \code{provoc}.}

\item{type}{"used" for the total number of times a mutation was used in the analysis and "coverage" for information about the coverage. Default \code{"coverage"}.}

\item{fun}{The function used to summarise coverage (when type = "coverage"). Default \code{mean}.}

\item{...}{Further arguments passed on to \code{fun}}

\item{main}{A main title for the plot.}

\item{col}{Colours to be used in plotting. Default \code{hcl.colors(n = 21, palette = "Dark Mint", rev = TRUE)}. Single-hue sequential colour palettes recommended.}
}
\description{
Visualize how lineage definitions were used in analysis
}
\examples{

data(Baaijens)
b2 <- Baaijens [Baaijens$sra \%in\% unique(Baaijens$sra)[1:30], ]
b2$mutations <- parse_mutations(b2$label)

lineage_defs <- astronomize() |>
    filter_lineages(c("B.1.617.1", "B.1.617.2", "B.1.617.2+K417N",
        "B.1.427", "B.1.429", "B.1.1.7"))
res <- provoc(
    formula = count / coverage ~ .,
    lineage_defs = lineage_defs,
    data = b2, 
    by = "sra",
    bootstrap_samples = 0)
plot_actual_defs(res, type = "coverage", fun = max, na.rm = TRUE)
}
