% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_diagnostics.R
\name{plot_lineage_defs}
\alias{plot_lineage_defs}
\title{Plot a lineage definition matrix}
\usage{
plot_lineage_defs(
  lineage_defs,
  col = colorRampPalette(colors = c("white", "dodgerblue4")),
  main = NULL
)
}
\arguments{
\item{lineage_defs}{Lineage definition matrix, such as those produced by \code{astronomize()} and \code{usher_barcodes()}.Alternatively the output of \code{provoc()}, from which the lineage definition matrix will be extracted.}

\item{col}{A function that takes argument \code{n} and returns a vector of colours of length \code{n}, such as \code{colorRampPallette(c("white", "dodgerblue4"))}}

\item{main}{a main title for the plot.}
}
\description{
Plot a lineage definition matrix
}
\details{
Every second column in the display has a slight grey background to make it easier to read which mutation is in which lineage.

The final column summarises the number of mutations that are unique to a given variant, with the darker colour representing a variant with more unique mutations. The number in the brackets gives the exact number of unique mutations.
}
\examples{
# Choose lineages that I know have a lot of shared mutations
lineage_defs <- astronomize() |>
    filter_lineages(c("B.1.1.7", "B.1.617.1", "B.1.617.2",
        "B.1.617.2+K417N", "B.1.617.3", "XBB.1"))
# XBB.1 has 1 shared mutation and 2 unique mutations
# None of B.1.1.7's mutations are shared
plot_lineage_defs(lineage_defs)

}
