% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{plot_lineages}
\alias{plot_lineages}
\title{plot the similarities of lineage}
\usage{
plot_lineages(provoc_obj, type = "Jaccard_similarity", labels = TRUE)
}
\description{
plot the similarities of lineage
}
