% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{plot_resids}
\alias{plot_resids}
\title{Plot the residuals, by variant}
\usage{
plot_resids(provoc_obj, type = "deviance", by_variant = TRUE)
}
\arguments{
\item{provoc_obj}{Result of fitting provoc().}

\item{type}{Deviance or raw residuals.}
}
\description{
Plot the residuals, by variant
}
