% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{plot_variants}
\alias{plot_variants}
\title{plot the similarities of variants}
\usage{
plot_variants(provoc_obj, type = "Jaccard_similarity", labels = TRUE)
}
\description{
plot the similarities of variants
}
