% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R, R/parse_mutations.R
\name{pos_from_aa}
\alias{pos_from_aa}
\title{Find position from aa}
\usage{
pos_from_aa(aa)

pos_from_aa(aa)
}
\arguments{
\item{aa}{A vector of mutations. aa:orf1a:I300V, ins:28215:3, del:27378:25, or C703T}
}
\description{
Find position from aa

Find position from aa
}
