% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_methods.R
\name{predict.provoc}
\alias{predict.provoc}
\title{Predict using Proportions of Variants of Concern}
\usage{
\method{predict}{provoc}(
  provoc_obj,
  newdata = NULL,
  type = NULL,
  dispersion = NULL,
  terms = NULL
)
}
\arguments{
\item{provoc_obj}{Named list with \code{proportions} and \code{variant_matrix}.}

\item{newdata}{Not yet implemented.}

\item{type}{Not yet implemented.}

\item{dispersion}{Not yet implemented.}

\item{terms}{Not yet implemented.}
}
\value{
Predicted values in the same order as the input data.
}
\description{
Takes a named list with an estimate of the proportions and the
associated variant matrix, performs matrix multiplication to
predict outcomes, and returns results in the same order as the original data.
}
\examples{
predicted_results <- predict(provoc_obj)
}
