% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{prepare_and_fuse_data}
\alias{prepare_and_fuse_data}
\title{Prepare and Fuse Data}
\usage{
prepare_and_fuse_data(data, lineage_defs, by, verbose)
}
\arguments{
\item{data}{A data frame containing the variables in the model.}

\item{lineage_defs}{A matrix of mutation definitions.}

\item{by}{An optional string specifying the column name to group the data by.}

\item{verbose}{}
}
\value{
A list containing the fused data frame and the grouped data as a list (if applicable).
}
\description{
Prepares the data based on the grouping variable and applies the \code{fuse} function.
}
\examples{
# This function is internally used and not typically called by the user.
}
