% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{print.provoc}
\alias{print.provoc}
\title{Print the results of lineage abundance estimation}
\usage{
\method{print}{provoc}(provoc_obj, n = 6)
}
\description{
Print the results of lineage abundance estimation
}
