% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{print.summary.provoc}
\alias{print.summary.provoc}
\title{Print the summary of a provoc object}
\usage{
\method{print}{summary.provoc}(summary.provoc)
}
\description{
Print the summary of a provoc object
}
