% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{process_lineage_defs}
\alias{process_lineage_defs}
\title{Process Mutation Definitions}
\usage{
process_lineage_defs(lineage_defs)
}
\arguments{
\item{lineage_defs}{A matrix of mutation definitions or NULL to use the default generated by \code{astronomize()}.}
}
\value{
A matrix of mutation definitions ready for analysis.
}
\description{
Handles mutation definitions by using the provided matrix or generating it using \code{astronomize()}.
}
\examples{
# This function is internally used and not typically called by the user.
}
