% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{process_mutation_defs}
\alias{process_mutation_defs}
\title{Process Mutation Definitions}
\usage{
process_mutation_defs(mutation_defs)
}
\arguments{
\item{mutation_defs}{A matrix of mutation definitions or NULL to use the default generated by \code{astronomize()}.}
}
\value{
A matrix of mutation definitions ready for analysis.
}
\description{
Handles mutation definitions by using the provided matrix or generating it using \code{astronomize()}.
}
\examples{
# This function is internally used and not typically called by the user.
}
