% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{re.calc.T.para}
\alias{re.calc.T.para}
\title{calculate the parameters of time}
\usage{
re.calc.T.para(L = sp)
}
\arguments{
\item{L}{sp}
}
\value{
sp
}
\description{
Be careful. By default, the "time" option uses a relative T0 from the first
spectrum of each acquisition and the "date" option uses the actual date and
time of each spectrum.
Using the acq_T0 column with the "time" option allows different acquisitions
to be sequenced using the T0 of the specified acquisition. Using the acq_T0
column with the "date" option allows you to overlap acquisitions on the T0
of the specified acquisition.
}
\details{
The delta_T column is used to add the specified time (in seconds) to the
acquisition.
}
\examples{
# The first time configuration :
#
# sp <- import.meta("meta_1")
# sp <- re.calc.T.para(sp)
#
# The second time configuration :
#
# sp <- import.meta("meta_2")
# sp <- re.init.T.para(sp)
# sp <- re.calc.T.para(sp)
}
