% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R
\name{re_findall}
\alias{re_findall}
\title{re_findall}
\usage{
re_findall(pat, s)
}
\arguments{
\item{pat:}{regex pattern}

\item{s:}{character, a single string}
}
\value{
character, vector of all matching substrings
}
\description{
Emulate behaviour of Python's re.findall() function. Lovingly stolen from \url{https://github.com/PoonLab/gromstole}.
}
