% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{remove_identical_variants}
\alias{remove_identical_variants}
\title{Remove Identical variants}
\usage{
remove_identical_variants(fused_df, annihilate)
}
\arguments{
\item{fused_df}{The fused data frame from the \code{fuse()} function}

\item{annihilate}{if TRUE will remove all duplicate variants, if FALSE will warn user if there is duplicate variants}
}
\value{
A data frame with no mutations that have a duplicate combination of variants
}
\description{
To ensure the predictor matrix is singular, if annihilate is TRUE the function returns all the mutations with a unique combination of variants. If FALSE it warns user of duplicate variants
}
\examples{
# This function is internally used and not typically called by the user.
}
