% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_methods.R
\name{residuals.provoc}
\alias{residuals.provoc}
\title{Calculate the residuals of a provoc object}
\usage{
\method{residuals}{provoc}(provoc_obj, type = "deviance")
}
\arguments{
\item{provoc_obj}{The result of provoc()}

\item{type}{"deviance" or "raw"}
}
\description{
Calculate the residuals of a provoc object
}
