% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_optim.R
\name{rho_initializer}
\alias{rho_initializer}
\title{Initialize a vector of proportions for the lineages of concern, prioritizing current (March 2022) most probable VoCs.}
\usage{
rho_initializer(lineage_defs)
}
\arguments{
\item{lineage_defs}{The lineage matrix to be used in the study; only used for the rownames (which should be VOCs in the expected format)}
}
\value{
lineage_defs a vector with the same length as the number of rows of lineage_defs, such that the values sum to less than one and each value is between 0 and 1
}
\description{
Initialize a vector of proportions for the lineages of concern, prioritizing current (March 2022) most probable VoCs.
}
