% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_coco.R
\name{simulate_coco}
\alias{simulate_coco}
\title{Simulate counts and coverage from a variant matrix}
\usage{
simulate_coco(varmat, rel_counts = NULL, censoring = NULL)
}
\arguments{
\item{varmat}{The variant matrix, with properly named rows (VOCs) and columns (mutations)}

\item{rel_counts}{(Optional) The relative counts of each VOC (will be censored by coverage). Must be same length as nrow(varmat).}

\item{censoring}{(Optional) The proportion of observations for each mutation. Must be same length as ncol(varmat).}
}
\value{
A data frame with counts and coverage of the sampled mutations, ready to be used in \code{copt_binom()}.
}
\description{
Simulate counts and coverage from a variant matrix
}
\details{
By default, uses a negative binomial distribution for the counts and uniform censoring for the coverage.
}
