% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{summarise_lineages}
\alias{summarise_lineages}
\title{Summarise the similarities in lineage matrices}
\usage{
summarise_lineages(provoc_obj)
}
\arguments{
\item{provoc_obj}{The result of provoc(), or an object coerced via as.provoc().}
}
\description{
Prints the most useful diagnostics to the screen, invisibly returning them as a list.
}
