% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{summary.provoc}
\alias{summary.provoc}
\title{Summarise results of model fitting}
\usage{
\method{summary}{provoc}(provoc_obj)
}
\description{
Prints the most useful diagnostics to the screen, invisibly returning them as a list.
}
