% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copt_binom.R
\name{to_feasible}
\alias{to_feasible}
\title{Transform a vector to the interior of the feasible region to be used as inits for constrOptim}
\usage{
to_feasible(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A vector with the same length of x such that the sum of the values is larger than 0 but less than 1 and all values are between 0 and 1
}
\description{
Transform a vector to the interior of the feasible region to be used as inits for constrOptim
}
