% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_diagnostics.R
\name{total_lineage_defs}
\alias{total_lineage_defs}
\title{Adds a list of lineage definition matrices together.}
\usage{
total_lineage_defs(ldef_list, summarise = "add")
}
\arguments{
\item{ldef_list}{A list of lineage definition matrices, such as those extracted by \code{get_actual_defs}.}
}
\description{
Entries represent the total number of times that lineage/mutation combination were used during fitting. Useful for checking whether a mutation was not present in a given sample.
}
