% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs for provoc}
\usage{
validate_inputs(formula, data)
}
\arguments{
\item{formula}{[stats]{formula}, specifying the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
None
}
\description{
Checks if the provided formula and data frame are valid for analysis.
}
\examples{
# This function is internally used and not typically called by the user.
}
