% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{variants_similarity}
\alias{variants_similarity}
\title{Finds and prints all similarities among variants}
\usage{
variants_similarity(data, is_varmat)
}
\arguments{
\item{data}{A dataframe either before or after it has been fused with varmat}

\item{is_varmat}{TRUE if data is a varaint matrix, FALSE if data is a fused dataframe}
}
\value{
none
}
\description{
Finds and prints all similarities among variants
}
