% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{varmat_from_data}
\alias{varmat_from_data}
\title{Use mutations found in the data to determine the variant matrix}
\usage{
varmat_from_data(
  type = NULL,
  pos = NULL,
  alt = NULL,
  aa = NULL,
  data = NULL,
  max_n = NULL,
  top_quantile = NULL
)
}
\arguments{
\item{type}{A vector containing either +, -, or ~, representing insertions, deletions, and mutations, respectively. Can also be i, d, or m.}

\item{pos}{A vector containing the 1-indexed positions.}

\item{alt}{A vector containing the alternative (nucleotide(s) for mutations and insertions or the number of deletions).}

\item{aa}{The vector of mutations in amino acid format (like those create by \code{parse_mutation}). TODO: Not yet implemented.}

\item{data}{A data frame containing columns labelled "type", "pos", and "alt", each with entries in the expected format. TODO: Not yet implemented.}

\item{max_n}{The maximum number of mutations from a given lineage to retain.}

\item{top_quantile}{Only take mutations that are in the top \code{top_quantile} quantile. E.g. 0.95 gives the mutations that are observed in more than 95\% of the sequences of that lineage.}
}
\value{
A variant matrix (rownames are variants, colnames are mutations, entry i,j is 1 if variant i contains mutation j, 0 otherwise).
}
\description{
Uses the included GenBank mutation list (\code{?mutations_by_lineage}) to determine a variant matrix that is guaranteed to have the mutations from your data (assuming you provide the correct mutation name format).
}
