% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{varmat_from_list}
\alias{varmat_from_list}
\title{Create a variant matrix from a list of variants with their mutations}
\usage{
varmat_from_list(variant_list)
}
\arguments{
\item{variant_list}{A named list of vectors of mutations. Mutation names should be in the same format as those in your data, otherwise post-processing will be required.}
}
\value{
A variant matrix (rownames are variants, colnames are mutations, entry i,j is 1 if variant i contains mutation j, 0 otherwise).
}
\description{
The input should be a named list where the names represent lineages and the values are vectors of mutations. The output will be a valid variant matrix.
}
