% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmat_helpers.R
\name{varmat_from_variants}
\alias{varmat_from_variants}
\title{Given a vector of lineage names, uses GenBank data to determine the variant matrix.}
\usage{
varmat_from_variants(variants, max_n = NULL, top_quantile = NULL)
}
\arguments{
\item{variants}{A character vector of variant names in the same format as the}

\item{max_n}{The maximum number of mutations from a given lineage to retain.}

\item{top_quantile}{Only take mutations that are in the top \code{top_quantile} quantile. E.g. 0.05 gives the mutations in more than 95\% of the sequences of that lineage.}
}
\value{
A variant matrix (rownames are variants, colnames are mutations, entry i,j is 1 if variant i contains mutation j, 0 otherwise).
}
\description{
Requires one of a few potentially arbitrary choices: Use the \code{max_n} most common mutations for a given variant or use the mutations in the top \code{top_quantile} quantile.
}
\details{
See \code{?mutations_by_lineage} for more information about the GenBank data used and see \code{unique(mutations_by_lineage$lineage)} to see available lineage names.
}
