% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomize.R
\name{astronomize}
\alias{astronomize}
\title{Extract mutation list from a directory of constellation files.}
\usage{
astronomize(path = NULL)
}
\arguments{
\item{path}{Path to the constellations folder in the cov-lineages/constellations repository. If NULL, built-in definitions are used. There is no need to clone the repo unless you want updated definitions.}
}
\value{
A lineage definition matrix for use with provoc.
}
\description{
"Constellations" are files produced from \url{https://github.com/cov-lineages/constellations}, which represent the cov-lineage team's best knowledge about which mutations define a lineage. See details.
}
\details{
The constellations repo is no longer updated to new variants.

Code is adapted from \code{scripts/estimate-freqs.R} in \url{https://github.com/PoonLab/gromstole}.

From the repo, a constellation "a collection of mutations which are functionally meaningful, but which may arise independently a number of times".

Updated as of version 0.5.0: constellation files are hardcoded into \code{provoc}, there is no longer a need to clone the repository.

If a path is provided, the function uses that repository. This is useful if, say, you clone a historical version of the repo to ensure that you'r eworking with definitions that were known at the time.
}
\examples{
# After cloning the constellations repo
lineage_defs <- astronomize(path = "../constellations")
dim(lineage_defs)
lineage_defs <- filter_lineages(lineage_defs, c("B.1.1.7", "B.1.617.2"))
dim(lineage_defs) # rows and columns have changed

}
\seealso{
\code{\link[=filter_lineages]{filter_lineages()}} to subset according to a vector of variants.
}
