% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{autoplot.provoc}
\alias{autoplot.provoc}
\title{Plot a provoc object using ggplot2}
\usage{
\method{autoplot}{provoc}(provoc_obj, date_col = NULL)
}
\arguments{
\item{provoc_obj}{Resulting object of class provoc to be used in plotting.}

\item{date_col}{Optional - if there's a date column, the results are plotted over time. This can be problematic if there are multiple samples at each time point.}
}
\description{
Plots the results of estimating wastewater prevalence of SARS-CoV-2. Optionally plots the results over time if given a date column.
}
\examples{
data(Baaijens)
b10 <- Baaijens[Baaijens$sra \%in\% unique(Baijens$sra)[1:10], ]
b10$mutations <- parse_mutations(b10$label)
res <- provoc(count / coverage ~ B.1.1.7 + B.1.617.2,
    data = b10, by = "sra")
res$date <- lubridate::ymd(res$date)
autoplot(res, date_col = "date")

}
