% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R, R/provoc_methods.R
\name{autoplot.provoc}
\alias{autoplot.provoc}
\title{Plot a provoc object using ggplot2}
\usage{
\method{autoplot}{provoc}(provoc_obj)

\method{autoplot}{provoc}(provoc_obj)
}
\arguments{
\item{provoc_obj}{A \code{provoc} class object containing variant proportions and, optionally, multiple sample groups.}

\item{date_col}{Optional - if there's a date column, the results are plotted over time. This can be problematic if there are multiple samples at each time point.}
}
\value{
A ggplot object visualizing the variant proportions. If multiple samples are present,
facets the plot for each sample.
}
\description{
Plots the results of estimating wastewater prevalence of SARS-CoV-2. Optionally plots the results over time if given a date column.

Generates a ggplot2 bar plot for the proportions of variants of concern from a \code{provoc} object.
Dynamically adjusts to handle both single and multiple samples within the same object.
}
\examples{
autoplot.provoc(res)
}
