% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_variants.R
\name{get_canada_variants}
\alias{get_canada_variants}
\title{Download weekly breakdown of SARS-CoV-2 variants in Canada}
\usage{
get_canada_variants(path = "data/canada_variants.csv")
}
\arguments{
\item{path}{Where to read the file (or store it if not already downloaded). Relative paths are fine. I suggest using the \code{here} package to get full path names so this function can do better error checking.}
}
\value{
A data frame with weekly variant numbers
}
\description{
Download weekly breakdown of SARS-CoV-2 variants in Canada
}
