% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{get_mutation_defs}
\alias{get_mutation_defs}
\title{Extract the mutation definitions used to fit the model}
\usage{
get_mutation_defs(provoc_obj)
}
\arguments{
\item{provoc_obj}{Object of class provoc to be used to extract the mutation definitions.}
}
\description{
Prints the most useful diagnostics to the screen, invisibly returning them as a list.
}
