% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{get_res}
\alias{get_res}
\title{Extract just the results of lineage estimation}
\usage{
get_res(provoc_obj)
}
\arguments{
\item{provoc_obj}{Object of class provoc to be used to extract the results.}
}
\description{
Prints the most useful diagnostics to the screen, invisibly returning them as a list.
}
