% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{import.h5}
\alias{import.h5}
\title{Import all file.h5 in the h5 folder of working directory.}
\usage{
import.h5(
  wdir = getwd(),
  pk_param = NULL,
  ctrl_peak = FALSE,
  baseline_correction = TRUE,
  skip = FALSE
)
}
\arguments{
\item{wdir}{the working directory}

\item{pk_param}{a set of parameters for importation and detection peak}

\item{ctrl_peak}{logical. An option for create a visual detect peak checking plot}

\item{baseline_correction}{logical. If TRUE, the baseline is correted}

\item{skip, }{numeric. The number of non-imported spectra (starting with the first)}
}
\value{
sp
}
\description{
For use this functions, you must have a folder name "h5" whith acquisition inside.
}
\examples{
# wd <- "C:/Users/huguenin/Documents/R/provoc test/data test/miscalenous"
#
# /!\ Note : Your datas are store like that :
# "wd/h5/00_file_PTR_ToF_MS.h5"
# "wd/h5/01_file_PTR_ToF_MS.h5"
# "wd/h5/02_file_PTR_ToF_MS.h5"
#
# setwd(wd)
# sp <- import.h5(wd)
#
# For the parameters :
# pk_param =  c(NULL,"very hight", "hight", "medium", "low")
# This determine the sensibility of your peak target.
# list(method = "MAD", halfWindowSize = 5, SNR = 10, smooth = 6) #NULL
# list(method = "MAD", halfWindowSize = 2, SNR = 10, smooth = 6) # very hight
# list(method = "MAD", halfWindowSize = 2, SNR = 40, smooth = 6) # hight
# list(method = "MAD", halfWindowSize = 5, SNR = 40, smooth = 6) # medium
# list(method = "MAD", halfWindowSize = 10, SNR = 60, smooth = 6) # low
#
# enougth, you can directly choose parameters :
# pk_param = list(method = "MAD", halfWindowSize = 2, SNR = 40, smooth = 6)
# method = "MAD" or "SuperSmoother"
# halfwindSize, SNR and smooth are integers
}
