% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_00_general.R
\name{kinetic.plot}
\alias{kinetic.plot}
\title{Title}
\usage{
kinetic.plot(
  M_num = M.Z(c(59, 137)),
  each_mass = TRUE,
  group = FALSE,
  graph_type = "dy",
  L = sp,
  Y_exp = FALSE,
  time_format = "date"
)
}
\arguments{
\item{time_format}{}
}
\value{

}
\description{
Title
}
