% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_general.R
\name{peak.ctrl}
\alias{peak.ctrl}
\title{print a graph with MS and peaks detected.}
\usage{
peak.ctrl(peak = 137, L = sp, suffixe = "")
}
\arguments{
\item{peak}{a number corresponding to a mass to be checked}

\item{L}{sp}

\item{suffixe}{a precision about this peak or this serie}
}
\value{
a plot
}
\description{
The graph will be centered on peak. It is not necessary to be precise for this number.
}
\examples{
# for(i in 50:210) peak.ctrl(i)
}
