% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R, R/provoc_methods.R
\name{plot.provoc}
\alias{plot.provoc}
\title{Plot the results of model fitting}
\usage{
\method{plot}{provoc}(provoc_obj)

\method{plot}{provoc}(provoc_obj)
}
\arguments{
\item{provoc_obj}{A \code{provoc} class object containing VoC proportion data and, optionally, multiple sample groups.}

\item{plot_type}{Currently only "barplot" is implemented. Residual plots and other diagnostics are works in progress.}
}
\value{
Generates a bar plot or series of bar plots visualizing the proportion of each variant.
}
\description{
Generates bar plots for the proportions of variants of concern (VoC) contained within a \code{provoc} object.
It can handle both single and multiple samples within the same \code{provoc} object.
}
\examples{
plot.provoc(provoc_obj)

}
