% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_diagnostics.R
\name{plot_lineage_defs2}
\alias{plot_lineage_defs2}
\title{Compare two lineage definition matrices}
\usage{
plot_lineage_defs2(left_def, right_def, col = NULL, n_colours = 20, ...)
}
\arguments{
\item{left_def, right_def}{Lineage definition matrice, such as those produced by \code{astronomize()} and \code{usher_barcodes()}.}

\item{col}{A named vector of colours, or a function with a single argument (n) that creates a colour scheme, such as \code{colorRampPalette()}. See Details and Examples.}

\item{n_colours}{The number of colours to be used in the function that creates a diverging colour scheme.}

\item{...}{Arguments passed on to \code{heatmap()}, such as \code{main} and \code{mar}.}
}
\description{
Compare two lineage definition matrices
}
\details{
A darker colour means that the mutation is present in that lineage, whereas a lighter colour means it's absent.

Purple represents mutation/lineage combinations that are only present in the left definitions, and orange means it's only in the right.

The center square represents mutations and lineages that are present in both lineage definitions, with white indicating that the mutation is not present for that particular lineage and black indicating that it is present.

The colours can be a named vector, with names "NA_col", "left_0", "left_1", "NA_col", "right_0", "right_1", "neither", "left_1", "right_1", and "both". Alternatively, a diverging colour scale can be provided (the colour scale must hae a single argument that determines the number of colours).
}
\examples{

lineages <- c("B.1.1.7", "P.1", "B.1.617.1", "B.1.617.2", "B.1.617.3")
left_def <- astronomize() |> filter_lineages(lineages[-1])
right_def <- usher_barcodes() |> filter_lineages(lineages[-2])
plot_lineage_defs2(left_def, right_def,
    main = "Constellations (without B.1.1.7) vs. Usher Barcodes (without P.1)")

my_cols <- colorRampPalette(c("dodgerblue", "white", "forestgreen"))
plot_lineage_defs2(left_def, right_def,
    main = "Constellations (without B.1.1.7) vs. Usher Barcodes (without P.1)",
    col = my_cols)

}
\seealso{
heatmap()
}
