% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_methods.R
\name{predict.provoc}
\alias{predict.provoc}
\title{Predict using Proportions of Variants of Concern}
\usage{
\method{predict}{provoc}(
  provoc_obj,
  b1,
  b2,
  newdata = NULL,
  type = NULL,
  dispersion = NULL,
  terms = NULL
)
}
\arguments{
\item{provoc_obj}{Named list with \code{proportions} and \code{variant_matrix}.}

\item{b1}{Name of the first sra sample group.}

\item{b2}{Name of the second sra sample group.}

\item{newdata}{Not yet implemented.}

\item{type}{Not yet implemented.}

\item{dispersion}{Not yet implemented.}

\item{terms}{Not yet implemented.}

\item{se.fit}{Not yet implemented.}

\item{na.action}{Not yet implemented.}
}
\value{
Predicted values in the same order as the input data.
}
\description{
Takes a named list with an estimate of the proportions and the
associated variant matrix, performs matrix multiplication to
predict outcomes, and returns results in the same order as the original data.
}
\examples{
library(provoc)
data("Baaijens")
Baaijens$mutation <- parse_mutations(Baaijens$label)
res <- provoc(formula = cbind(count, coverage) ~ B.1.1.7 + B.1.617.2, data = Baaijens)
b1 <- Baaijens[Baaijens$sra == unique(Baaijens$sra)[1], ]
b2 <- Baaijens[Baaijens$sra == unique(Baaijens$sra)[2], ]
predicted_values <- predict.provoc(res, b1, b2)
print(predicted_values)
}
