% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc_class.R
\name{print.provoc}
\alias{print.provoc}
\title{Print the results of lineage abundance estimation}
\usage{
\method{print}{provoc}(provoc_obj, n = 6)
}
\arguments{
\item{provoc_obj}{The resulting object of class provoc to be used in printing.}

\item{n}{The number of rows of the results dataframe to print}
}
\description{
Prints the most useful diagnostics to the screen, invisibly returning them as a list.
}
