% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{process_optim}
\alias{process_optim}
\title{Process Optimization}
\usage{
process_optim(grouped_data, mutation_defs, by, bootstrap_samples, verbose)
}
\arguments{
\item{grouped_data}{A list containing data frames for each group to be processed.}

\item{mutation_defs}{A matrix of mutation definitions.}

\item{by}{An optional string specifying the column name to group the data by.}

\item{bootstrap_samples}{The number of bootstrap samples to use.}

\item{verbose}{TRUE to print detailed messages.}
}
\value{
A list of results for each group, including point estimates and convergence information.
}
\description{
Processes each group or the entire dataset through \code{provoc_optim} and collects results.
}
\examples{
# This function is internally used and not typically called by the user.
}
