% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{provoc}
\alias{provoc}
\title{Proportions of Variants of Concern (provoc) Analysis}
\usage{
provoc(
  formula,
  data,
  mutation_defs = NULL,
  by = NULL,
  update_interval = 20,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A formula for the binomial model, like cbind(count, coverage) ~ .}

\item{data}{Data frame containing count, coverage, and lineage columns.}

\item{mutation_defs}{Optional mutation definitions; if NULL, uses astronomize().}

\item{by}{Column name to group and process data.}

\item{update_interval}{Interval for progress messages (0 to suppress).}

\item{verbose}{TRUE to print detailed messages.}
}
\value{
Returns an object of class 'provoc' with results from applying \code{provoc_optim} to the input data. The object contains the following attributes:
\itemize{
\item proportions: Estimated proportions vector for each variant of concern.
\item variant_matrix: Mutation definitions used for analysis, provided \code{mutation_defs} or default \code{astronomize()}.
}

Outputs necessary information for subsequent analysis, including the use of the \code{predict.provoc()}.
}
\description{
Applies provoc_optim to analyze COVID-19 Variant of Concern proportions.
It allows flexible lineage and mutation definitions.
}
\examples{
library(provoc)
# Load a dataset
data("Baaijens")
# Prepare the dataset
Baaijens$mutation <- parse_mutations(Baaijens$label)

# Analyze the dataset using the default mutation definitions
res <- provoc(formula = cbind(count, coverage) ~ B.1.1.7 + B.1.617.2,
              data = Baaijens, by = "sample_id")

# Check for analysis convergence
print(convergence(res))

# Use the results for prediction
predicted_values <- predict.provoc(res)

}
