% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoc.R
\name{provoc}
\alias{provoc}
\title{Proportions of Variants of Concern}
\usage{
provoc(fused, method = c("optim", "runjags"))
}
\arguments{
\item{fused}{The fused data frame of coco and varmat}

\item{method}{Optimization or Bayesian?}
}
\value{
Results of the estimation. Currently has different output depending on method, but will hopefully be unified in a later version.
}
\description{
Un-fuses coco and varmat and applies the appropriate estimation technique. Applies to each unique value of
}
