% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copt_binom.R
\name{rho_initializer}
\alias{rho_initializer}
\title{Initialize a vector of proportions for the variants of concern, prioritizing current (March 2022) most probable VoCs.}
\usage{
rho_initializer(varmat)
}
\arguments{
\item{varmat}{The variant matrix to be used in the study; only used for the rownames (which should be VOCs in the expected format)}
}
\value{
varmat a vector with the same length as the number of rows of varmat, such that the values sum to less than one and each value is between 0 and 1
}
\description{
Initialize a vector of proportions for the variants of concern, prioritizing current (March 2022) most probable VoCs.
}
