% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{simplify_similarity}
\alias{simplify_similarity}
\title{Simplifies variant similarity matrices for easier interpretation}
\usage{
simplify_similarity(similarities, almost)
}
\arguments{
\item{similarities}{Result of \code{variants_similarity()}}

\item{almost}{Degree of similarity.}
}
\description{
Simplifies variant similarity matrices for easier interpretation
}
