% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lineage_defs.R
\name{simulate_lineage_defs}
\alias{simulate_lineage_defs}
\title{Simulate a small lineage matrix for testing}
\usage{
simulate_lineage_defs(
  lins = c("BA.1", "BA.2", "B.1.1.529"),
  mutations = c("m3037T", "m22599A", "d221943", "i22205GAGCCAGAA", "m24469A", "d219879",
    "m241T")
)
}
\arguments{
\item{lins}{Names of lineages}

\item{mutations}{Names of mutations}
}
\value{
A properly structured lineage definitions matrix (with rownames and colnames).
}
\description{
Creates a matrix with the specified VoCs and mutation names, but randomly applies mutations. A future version of this function may be useful for simulating mutations in a more meaningful way, but for now it's only useful for testing purposes.
}
