% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{variants_similarity}
\alias{variants_similarity}
\title{Finds and prints all similarities among variants}
\usage{
variants_similarity(data, simplify = FALSE, almost = 1)
}
\arguments{
\item{data}{A variant matrix}
}
\value{
A list of length 4 containing information on which variants differ by one,
the Jaccard similarity between variants, which variants are subsets and almost subsets
of each other. in is_subset and is_almost_subset a value is true if the variant of the
column name is a subset/almost a subset of the variant of the row name.
}
\description{
Finds and prints all similarities among variants
}
